+++ Java/AdventOfCodeJava/.idea/.gitignore
@@ -6,3 +6,9 @@
 # Datasource local storage ignored files
 /dataSources/
 /dataSources.local.xml
+!/sonarlint/issuestore/
+!/sonarlint/
+!/compiler.xml
+!/git_toolbox_prj.xml
+!/jarRepositories.xml
+!/workspace.xml
