+++ Java/AdventOfCodeJava/src/main/java/org/jmitchell238/aoc/Main.java
@@ -1,7 +1,7 @@
 package org.jmitchell238.aoc;
 
 import java.io.FileNotFoundException;
-import org.jmitchell238.aoc.aoc2023.Day1;
+import org.jmitchell238.aoc.aoc2023.Day01.Day01;
 
 public class Main {
 
@@ -9,6 +9,6 @@ public class Main {
     System.out.println("Hello Advent of Code 2023!");
 
     // Day 1
-    Day1.Day1();
+    Day01.Day01();
   }
 }
