+++ Java/AdventOfCodeJava/src/main/java/org/jmitchell238/aoc/aoc2023/Day01/Day01.java
@@ -1,61 +1,4 @@
-/**
- * --- Day 1: Trebuchet?! ---
- *
- * Something is wrong with global snow production, and you've been selected to take a look. The Elves have even given you a map;
- * on it, they've used stars to mark the top fifty locations that are likely to be having problems.
- *
- * You've been doing this long enough to know that to restore snow operations, you need to check all fifty stars by December 25th.
- *
- * Collect stars by solving puzzles. Two puzzles will be made available on each day in the Advent calendar; the second puzzle is unlocked when
- * you complete the first. Each puzzle grants one star. Good luck!
- *
- * You try to ask why they can't just use a weather machine ("not powerful enough") and where they're even sending you ("the sky")
- * and why your map looks mostly blank ("you sure ask a lot of questions") and hang on did you just say the sky
- * ("of course, where do you think snow comes from") when you realize that the Elves are already loading you into a trebuchet
- * ("please hold still, we need to strap you in").
- *
- * As they're making the final adjustments, they discover that their calibration document (your puzzle input)
- * has been amended by a very young Elf who was apparently just excited to show off her art skills.
- * Consequently, the Elves are having trouble reading the values on the document.
- *
- * The newly-improved calibration document consists of lines of text; each line originally contained a specific calibration value that the Elves now
- * need to recover. On each line, the calibration value can be found by combining the first digit and the last digit (in that order)
- * to form a single two-digit number.
- *
- * For example:
- *
- * 1abc2
- * pqr3stu8vwx
- * a1b2c3d4e5f
- * treb7uchet
- *
- * In this example, the calibration values of these four lines are 12, 38, 15, and 77. Adding these together produces 142.
- *
- * Consider your entire calibration document. What is the sum of all of the calibration values?
- */
-
-/*
-  * --- Part Two ---
-  *
-  * Your calculation isn't quite right. It looks like some of the digits are actually spelled out with letters:
-  * one, two, three, four, five, six, seven, eight, and nine also count as valid "digits".
-  *
-  * Equipped with this new information, you now need to find the real first and last digit on each line. For example:
-  *
-  * two1nine
-  * eightwothree
-  * abcone2threexyz
-  * xtwone3four
-  * 4nineeightseven2
-  * zoneight234
-  * 7pqrstsixteen
-  *
-  * In this example, the calibration values are 29, 83, 13, 24, 42, 14, and 76. Adding these together produces 281.
-  *
-  * What is the sum of all of the calibration values?
- */
-
-package org.jmitchell238.aoc.aoc2023;
+package org.jmitchell238.aoc.aoc2023.Day01;
 
 import java.io.File;
 import java.io.FileNotFoundException;
@@ -65,17 +8,17 @@ import java.util.Objects;
 import java.util.Scanner;
 
 
-public class Day1 {
+public class Day01 {
 
-  public static void Day1() throws FileNotFoundException {
+  public static void Day01() throws FileNotFoundException {
     System.out.println("Day 1: Trebuchet?!");
 
-    File day1Part1Input = new File("/Users/jamesm.mitchell/james-workspace/AdventOfCode/Java/AdventOfCodeJava/src/main/resources/aoc2023/day1_part1.txt");
+    File day1Part1Input = new File("/Users/jamesm.mitchell/james-workspace/AdventOfCode/Java/AdventOfCodeJava/src/main/resources/aoc2023/day1.txt");
     int partOneTotal = Part1(day1Part1Input);
 
     System.out.println("Part 1: Answer: " + partOneTotal);
 
-    File day1Part2Input = new File("/Users/jamesm.mitchell/james-workspace/AdventOfCode/Java/AdventOfCodeJava/src/main/resources/aoc2023/day1_part2.txt");
+    File day1Part2Input = new File("/Users/jamesm.mitchell/james-workspace/AdventOfCode/Java/AdventOfCodeJava/src/main/resources/aoc2023/day1.txt");
 
     int partTwoTotal = Part2(day1Part2Input);
     System.out.println("Part 2: Answer:" + partTwoTotal);
@@ -88,7 +31,7 @@ public class Day1 {
     try (Scanner scanner = new Scanner(day1Part1Input);) {
       while (scanner.hasNextLine()) {
         String line = scanner.nextLine();
-        StringBuilder numberString = new StringBuilder(line.replaceAll("[^0-9]", ""));
+        StringBuilder numberString = new StringBuilder(line.replaceAll("\\D", "")); // "\\D" is a regex for non-digits same as [^0-9]
 
 
         total = getTotal(numberString, total);
