+++ Java/AdventOfCodeJava/src/main/java/org/jmitchell238/aoc/aoc2023/Day02/Day02.java
@@ -1,0 +1,5 @@
+package org.jmitchell238.aoc.aoc2023.Day02;
+
+public class Day02 {
+
+}
