+++ Java/AdventOfCodeJava/src/main/resources/aoc2023/Day01/day1_part2_test.txt
