+++ Java/AdventOfCodeJava/src/test/java/org/jmitchell238/aoc/aoc2023/Day01/Day1Tests.java
@@ -1,0 +1,61 @@
+package org.jmitchell238.aoc.aoc2023.Day01;
+
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+
+import java.io.File;
+import java.io.FileNotFoundException;
+import org.junit.jupiter.api.Test;
+
+class Day01Tests {
+
+  @Test
+  void Day01Part1_CalledWithTestInput_Expect142() throws FileNotFoundException {
+    // Arrange
+    File day1Part1Input = new File(
+        "/Users/jamesm.mitchell/james-workspace/AdventOfCode/Java/AdventOfCodeJava/src/main/resources/aoc2023/day1_part1_test.txt");
+
+    // Assert
+    int expected = 142;
+
+    // Act
+    assertThat(Day01.Part1(day1Part1Input)).isEqualTo(expected);
+  }
+
+  @Test
+  void Day01Part1_CalledWithRealInput_Expect54940() throws FileNotFoundException {
+    // Arrange
+    File day1Part1Input = new File("/Users/jamesm.mitchell/james-workspace/AdventOfCode/Java/AdventOfCodeJava/src/main/resources/aoc2023/day1.txt");
+
+    // Assert
+    int expected = 54940;
+
+    // Act
+    assertThat(Day01.Part1(day1Part1Input)).isEqualTo(expected);
+  }
+
+  @Test
+  void Day01Part2_CalledWithTestInput_Expect281() throws FileNotFoundException {
+    // Arrange
+    File day1Part2Input = new File(
+        "/Users/jamesm.mitchell/james-workspace/AdventOfCode/Java/AdventOfCodeJava/src/main/resources/aoc2023/day1_part2_test.txt");
+
+    // Assert
+    int expected = 281;
+
+    // Act
+    assertThat(Day01.Part2(day1Part2Input)).isEqualTo(expected);
+  }
+
+  @Test
+  void Day01Part2_CalledWithRealInput_Expect54208() throws FileNotFoundException {
+    // Arrange
+    File day1Part2Input = new File("/Users/jamesm.mitchell/james-workspace/AdventOfCode/Java/AdventOfCodeJava/src/main/resources/aoc2023/day1.txt");
+
+    // Assert
+    int expected = 54208;
+
+    // Act
+    assertThat(Day01.Part2(day1Part2Input)).isEqualTo(expected);
+  }
+}
+
