+++ Java/AdventOfCodeJava/src/test/java/org/jmitchell238/aoc/aoc2023/Day02/Day02Tests.java
@@ -1,0 +1,17 @@
+package org.jmitchell238.aoc.aoc2023.Day02;
+
+import org.junit.jupiter.api.Test;
+
+class Day02Tests {
+
+  @Test
+  void Part1_CalledWithTestInput_Expect() {
+    // Arrange
+
+    // Assert
+
+    // Act
+
+  }
+
+}
